Moves ReceivePurchaserInfoListener.toReceiveCustomerInfoCallback
