+++ purchases/src/main/java/com/revenuecat/purchases/interfaces/ReceiveCustomerInfoCallback.kt
@@ -5,6 +5,7 @@
 package com.revenuecat.purchases.interfaces
 
 import com.revenuecat.purchases.CustomerInfo
+import com.revenuecat.purchases.PurchaserInfo
 import com.revenuecat.purchases.PurchasesError
 
 /**
@@ -22,4 +23,16 @@ interface ReceiveCustomerInfoCallback {
      * @param error A [PurchasesError] containing the reason for the failure of the call
      */
     fun onError(error: PurchasesError)
+}
+
+fun ReceivePurchaserInfoListener.toReceiveCustomerInfoCallback(): ReceiveCustomerInfoCallback {
+    return object : ReceiveCustomerInfoCallback {
+        override fun onReceived(customerInfo: CustomerInfo) {
+            this@toReceiveCustomerInfoCallback.onReceived(PurchaserInfo(customerInfo))
+        }
+
+        override fun onError(error: PurchasesError) {
+            this@toReceiveCustomerInfoCallback.onError(error)
+        }
+    }
 }
